requirements.txt
django==2.2
mysqlclient==2.1.0
ipython==7.10.1
PyMySQL==1.0.2
djangorestframework==3.13.1
markdown==3.3.6
django-filter==21.1
pyyaml==6.0